Changed Colors, added compiler to base class

+++ testers/LibftTester.py
@@ -4,7 +4,7 @@ import re
 import shutil
 import subprocess
 from collections import namedtuple
-from main import Colors, TestRunInfo
+from main import CT, TestRunInfo
 from testers.CommonTester import show_banner
 from testers.libft.ExecuteFsoares import ExecuteFsoares
 from testers.libft.ExecuteTripouille import ExecuteTripouille
@@ -13,50 +13,29 @@ logger = logging.getLogger()
 
 Tester = namedtuple("Test", "name executable")
 
-AVAILABLE_TESTERS = [#Tester('Tripouille', ExecuteTripouille),
-				   Tester('fsoares', ExecuteFsoares)]
+AVAILABLE_TESTERS = [  #Tester('Tripouille', ExecuteTripouille),
+    Tester('fsoares', ExecuteFsoares)
+]
 
 FUNCTIONS_UNDER_TEST = [
-	"isalpha",
-	"isdigit",
-	"isalnum",
-	"isascii",
-	"isprint",
-	"strlen",
-	"memset",
-	"bzero",
-	"memcpy",
-	"memmove",
-	"strlcpy",
-	"strlcat",
-	"toupper",
-	"tolower",
-	"strchr",
-	"strrchr",
-	"strncmp",
-	"memchr",
-	"memcmp",
-	"strnstr",
-	"atoi",
-	"calloc",
-	"strdup",
-	"substr",
-	"strjoin",
-	"strtrim",
-	"split",
-	"itoa",
-	"strmapi",
-	"striteri",
-	"putchar_fd",
-	"putstr_fd",
-	"putendl_fd",
-	"putnbr_fd"
+    "isalpha", "isdigit", "isalnum", "isascii", "isprint", "strlen", "memset", "bzero", "memcpy", "memmove", "strlcpy",
+    "strlcat", "toupper", "tolower", "strchr", "strrchr", "strncmp", "memchr", "memcmp", "strnstr", "atoi", "calloc",
+    "strdup", "substr", "strjoin", "strtrim", "split", "itoa", "strmapi", "striteri", "putchar_fd", "putstr_fd",
+    "putendl_fd", "putnbr_fd"
 ]
 
 
 def intersection(lst1, lst2):
 	lst3 = [value for value in lst1 if value in lst2]
 	return lst3
+
+
+def run_popen(command: str):
+	to_execute = command.split(" ")
+	print(f"\n{CT.CYAN}Executing: {CT.WHITE}{' '.join(to_execute)}{CT.NC}:")
+	process = subprocess.Popen(to_execute)
+	process.wait()
+	return process
 
 
 func_regex = re.compile(r"\w+\s+\*?ft_(\w+)\(.*")
@@ -79,7 +58,7 @@ class LibftTester():
 		compile_res = self.create_library()
 
 		if not compile_res:
-			return "Project failed to create library"
+			return
 
 		present = self.get_present()
 		to_execute = intersection(present, FUNCTIONS_UNDER_TEST)
@@ -105,7 +84,9 @@ class LibftTester():
 			return tx.execute()
 
 	def show_summary(self, norm: str, present, errors):
+
 		def get_norm_errors():
+
 			def get_fname(line):
 				return norm_func_regex.match(line).group(1)
 
@@ -116,18 +97,18 @@ class LibftTester():
 
 		norm_errors = get_norm_errors()
 		if norm_errors:
-			print(f"{Colors.LIGHT_RED}Norminette Errors:{Colors.NC}")
+			print(f"{CT.L_RED}Norminette Errors:{CT.NC}")
 			print(', '.join(norm_errors))
 
 		missing = [f for f in FUNCTIONS_UNDER_TEST if f not in present]
 		if missing:
-			print(f"\n{Colors.LIGHT_RED}Missing functions: {Colors.NC}{', '.join(missing)}")
+			print(f"\n{CT.L_RED}Missing functions: {CT.NC}{', '.join(missing)}")
 
 		if errors:
-			print(f"\n{Colors.LIGHT_RED}Failed tests: {Colors.NC}{', '.join(errors)}")
+			print(f"\n{CT.L_RED}Failed tests: {CT.NC}{', '.join(errors)}")
 
 		if not missing and not norm_errors and not errors:
-			print(f"🎉🥳{Colors.LIGHT_GREEN}All tests passed! Congratulations!{Colors.NC}🥳🎉")
+			print(f"🎉🥳 {CT.L_GREEN}All tests passed! Congratulations!{CT.NC} 🥳🎉")
 
 	def prepare_ex_files(self):
 		if os.path.exists(self.temp_dir):
@@ -145,63 +126,58 @@ class LibftTester():
 
 		result = subprocess.run(norm_exec, capture_output=True, text=True)
 
-		print(f"{Colors.CYAN}Executing: {Colors.WHITE}{' '.join(norm_exec)}{Colors.NC}:")
+		print(f"{CT.CYAN}\nExecuting: {CT.WHITE}{' '.join(norm_exec)}{CT.NC}:")
 		if result.returncode != 0:
-			print(f"{Colors.YELLOW}{result.stdout}{Colors.NC}")
+			print(f"{CT.YELLOW}{result.stdout}{CT.NC}")
 		else:
-			print(f"{Colors.GREEN}Norminette OK!{Colors.NC}")
+			print(f"{CT.GREEN}Norminette OK!{CT.NC}")
 
 		return result.stdout
 
 	def create_library(self):
 		logger.info(f"Calling make on directory {os.getcwd()}")
-		make_exec = ["make"]
 
-		print(f"{Colors.CYAN}Executing: {Colors.WHITE}{' '.join(make_exec)}{Colors.NC}:")
-		subprocess.run(["make", "fclean"], capture_output=True, text=True)
-		p = subprocess.run(make_exec, capture_output=True, text=True)
+		run_popen("make fclean")
+		make = run_popen("make")
 
-		if p.returncode == 0:
-			print(f"{Colors.GREEN}make: OK!{Colors.NC}")
+		if make.returncode == 0:
+			print(f"{CT.GREEN}make: OK!{CT.NC}")
 		else:
-			print(f"{Colors.LIGHT_RED}Problem creating library{Colors.NC}")
-			print(f"{Colors.YELLOW}{p.stdout}{Colors.NC}")
-			print(f"{Colors.RED}{p.stderr}{Colors.NC}")
+			print(f"{CT.L_RED}Problem creating library{CT.NC}")
 
-		return p.returncode == 0
+		return make.returncode == 0
 
 	def prepare_tests(self, testname):
-		try:
-			# delete destination folder if already present
-			temp_dir = os.path.join(self.temp_dir, testname)
-			if os.path.exists(temp_dir):
-				logger.info(f"Removing already present directory {temp_dir}")
-				shutil.rmtree(temp_dir)
+		# delete destination folder if already present
+		temp_dir = os.path.join(self.temp_dir, testname)
+		if os.path.exists(temp_dir):
+			logger.info(f"Removing already present directory {temp_dir}")
+			shutil.rmtree(temp_dir)
 
-			# copy test framework
-			tester_dir = os.path.join(self.tests_dir, testname)
-			logger.info(f"Copying from {tester_dir} to {temp_dir}")
-			shutil.copytree(tester_dir, temp_dir)
+		# copy test framework
+		tester_dir = os.path.join(self.tests_dir, testname)
+		logger.info(f"Copying from {tester_dir} to {temp_dir}")
+		shutil.copytree(tester_dir, temp_dir)
 
-			# copy compiled library
-			library = os.path.join(self.temp_dir, "libft.a")
-			logger.info(
-				f"Copying libft.a from {library} to {temp_dir}")
-			shutil.copy(library, temp_dir)
+		# copy compiled library
+		library = os.path.join(self.temp_dir, "libft.a")
+		if not os.path.exists(library):
+			raise Exception(f"{CT.L_RED}libft.a{CT.RED} was not created. " +
+			                "Please create it in the Makefile.")
+		logger.info(f"Copying libft.a from {library} to {temp_dir}")
+		shutil.copy(library, temp_dir)
 
-			# copy header
-			header = os.path.join(self.temp_dir, "libft.h")
-			logger.info(
-				f"Copying libft.h from {header} to {temp_dir}")
-			shutil.copy(header, temp_dir)
+		# copy header
+		header = os.path.join(self.temp_dir, "libft.h")
+		logger.info(f"Copying libft.h from {header} to {temp_dir}")
+		shutil.copy(header, temp_dir)
 
-			return True
-		except Exception as ex:
-			logger.info("Problem creating the files structure: ", ex)
-			return False
+		return True
 
 	def get_present(self):
 		header = os.path.join(self.temp_dir, "libft.h")
+		if not os.path.exists(header):
+			raise Exception(f"There is no {CT.L_RED}libft.h{CT.RED} present")
 		with open(header, "r") as h:
 			funcs_str = [line for line in h.readlines() if func_regex.match(line)]
 			return [func_regex.match(line).group(1) for line in funcs_str]
